+++ server/routes/client.js
@@ -1,14 +1,8 @@
 import { Router } from 'express';
-import cloudinary from 'cloudinary';
 import newClient from '../controllers/client';
-import { upload, dataUri } from '../controllers/upload';
-import httpStatus from '../HttpStatus/index';
+import { upload, cloudUpload }  from '../controllers/upload';
+import validateToken from '../middlewares/validateToken';
 
-cloudinary.v2.config({
-  cloud_name: process.env.CLOUD_NAME,
-  api_key: process.env.CLOUD_API_KEY,
-  api_secret: process.env.CLOUD_API_SECRET      
-});
 
 const router = Router();
 // TeleMed root route
@@ -20,31 +14,13 @@ router.get('/', (req, res) => {
 });
 
 router.post('/auth/signup/', newClient.createClient);
-router.get('/clients/', newClient.getAllClients);
 
-router.post('/clients/upload', upload, (req, res) =>{
-  if(req.file){
-    const file = dataUri(req).content;
-    return cloudinary.uploader.upload(file) 
-    .then(result => {
-      const img = result.url;
-      return res.status(httpStatus.OK).json({
-        message: 'Image uploaded to cloudinary successfully!',
-        data: {
-          img
-        }
-      })
-    }).catch( err => res.status(httpStatus.BAD_REQUEST).json({
-      message: 'Error uploading image',
-      data: {
-        err
-      }
-    }))
-  }
-}
-);
+router.get('/clients/', validateToken, newClient.getAllClients);
 
-router.post('/auth/signin', newClient.authenticateClient);
+// upload image to cloudinary
+router.post('/clients/upload', validateToken, upload, cloudUpload);
+
+router.post('/auth/signin', validateToken, newClient.authenticateClient);
 
 
 export default router;
